feat: :bug: Removing route bugs with next auth
I moved the main page.tsx into the playlist folder, so that next auth redirects to the '/playlist' route, as it is not redirecting to the main route: '/'

Authentication working
