+++ front/src/app/(auth-routes)/layout.tsx
@@ -9,9 +9,9 @@ interface PrivateLayoutProps {
 
 export default async function PrivateLayout({ children }: PrivateLayoutProps){
 	const session = await getServerSession(nextAuthOptions)
-
+	
 	if (session) {
-		redirect('/')
+		redirect('/playlist')
 	}
 
 	return <>{children}</>
